#1.7.4 OLED IIC达到38fps

+++ Controller/HARDWARE/091OLED.h
@@ -30,7 +30,7 @@
 #define Scroll_Cmd(cmd) OLED_WR_Byte(0x2e | cmd, OLED_CMD) // Æô¶¯¹ö¶¯
 #define OLED_ShowChinese(x, y, no) OLED_DrawBMP(x, y, 16, 16, no);
 void OLED_WR_Byte(u8 dat, u8 cmd);
-void OLED_Multi_WR_Byte(u8 *dat, u8 cmd, u8 len);
+void OLED_Multi_WR_Byte(u8 *dat, u8 cmd, u16 len);
 void OLED_Display_Cmd(u8 state);
 void OLED_Init(void);
 void OLED_Clear(void);
